# api/openapi.yaml
openapi: 3.1.0
info:
  title: LyraNest Community API
  version: 0.3.0
  description: Authenticated API for library discovery, playback, management, normalized sidecar lyrics, and persistent per-user listening data.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
...
  /api/v1/auth/register:
    post:
      security: []
      operationId: registerFirstAdministrator
      summary: Register the first and only bootstrap administrator
      description: Available only while the user database is empty. Registration closes atomically after the first successful request.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CredentialsRequest"
      responses:
        "201":
          description: Administrator created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          description: Registration is closed because a user already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/auth/login:
    post:
      security: []
      operationId: login
      summary: Create a persistent server-side session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CredentialsRequest"
      responses:
        "200":
          description: Login succeeded
          headers:
            Cache-Control:
              schema:
                type: string
                const: no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/auth/me:
    get:
      operationId: getCurrentUser
      summary: Return the authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/auth/logout:
    post:
      operationId: logout
      summary: Revoke the presented bearer token
      responses:
        "204":
          description: Session revoked
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/tracks:
    get:
      operationId: listTracks
      summary: List scanned tracks
      responses:
        "200":
          description: Scanned tracks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackListResponse"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/library/scan:
    post:
      operationId: scanLibrary
      summary: Rescan the configured music library
      description: Synchronously scans the library directory and replaces the in-memory track list when successful.
      responses:
        "200":
          description: Library scan completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryScanResponse"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/library/status:
    get:
      operationId: getLibraryStatus
      summary: Get music library scan status
      responses:
        "200":
          description: Current library status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryStatusResponse"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/tracks/{trackId}/lyrics:
    get:
      operationId: getTrackLyrics
      summary: Get normalized sidecar lyrics
      parameters:
        - name: trackId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Lyrics decoded and normalized to UTF-8
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LyricsResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "415":
          description: Unsupported lyrics encoding
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/v1/tracks/{trackId}/artwork:
    parameters:
      - name: trackId
        in: path
        required: true
        schema:
          type: string
    get:
      security: []
      operationId: getTrackArtwork
      summary: Get track artwork
      description: Returns the preferred sidecar artwork or falls back to embedded artwork from the audio file.
      responses:
        "200":
          $ref: "#/components/responses/Artwork"
        "304":
          description: Artwork has not changed for the supplied conditional request headers.
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    head:
      security: []
      operationId: headTrackArtwork
      summary: Inspect track artwork
      responses:
        "200":
          $ref: "#/components/responses/ArtworkHeaders"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/tracks/{trackId}/stream:
    parameters:
      - name: trackId
        in: path
        required: true
        schema:
          type: string
    get:
      security: []
      operationId: streamTrack
      summary: Stream the original audio file
      description: Supports byte ranges for direct playback without transcoding.
      parameters:
        - name: Range
          in: header
          required: false
          description: Requested byte range, such as bytes=0-1023.
          schema:
            type: string
        - name: If-Range
          in: header
          required: false
          description: Strong ETag or HTTP-date validator for Range. A matching validator returns 206; a stale validator ignores Range and returns the complete 200 response.
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/AudioStream"
        "206":
          $ref: "#/components/responses/PartialAudioStream"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    head:
      security: []
      operationId: headTrackStream
      summary: Inspect the original audio stream
      responses:
        "200":
          $ref: "#/components/responses/AudioStreamHeaders"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/me/collections:
    get:
      operationId: getCollections
      summary: Get the authenticated user's authoritative collection snapshot
      responses:
        "200":
          description: Server-authoritative favorites and playlists snapshot
          headers:
            Cache-Control:
              schema:
                type: string
                const: no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionsSnapshot"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/me/favorites/{trackId}:
    parameters:
      - $ref: "#/components/parameters/TrackID"
    put:
      operationId: addFavorite
      summary: Add a track to the authenticated user's favorites
      responses:
        "200":
          description: Updated collection snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionsSnapshot"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "413":
          $ref: "#/components/responses/UserDataLimit"
    delete:
      operationId: removeFavorite
      summary: Remove a track from the authenticated user's favorites
      responses:
        "200":
          description: Updated collection snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionsSnapshot"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/me/playlists:
    post:
      operationId: createPlaylist
      summary: Create an empty playlist
      description: Omit id for a server-generated playlist ID. When a client-generated id is supplied, the first request creates the playlist; a retry with the same normalized name returns the existing snapshot without creating a duplicate. Reusing an ID with a different normalized name returns a conflict.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaylistCreateRequest"
      responses:
        "201":
          description: Updated collection snapshot containing the new playlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionsSnapshot"
        "200":
          description: Existing collection snapshot returned for an idempotent retry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionsSnapshot"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "413":
          $ref: "#/components/responses/UserDataLimit"
  /api/v1/me/playlists/{playlistId}:
    parameters:
      - $ref: "#/components/parameters/PlaylistID"
    patch:
      operationId: updatePlaylist
      summary: Rename a playlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaylistRequest"
      responses:
        "200":
          description: Updated collection snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionsSnapshot"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deletePlaylist
      summary: Delete a playlist
      responses:
        "200":
          description: Updated collection snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionsSnapshot"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/me/playlists/{playlistId}/tracks/{trackId}:
    parameters:
      - $ref: "#/components/parameters/PlaylistID"
      - $ref: "#/components/parameters/TrackID"
    put:
      operationId: addPlaylistTrack
      summary: Add a track to a playlist
      responses:
        "200":
          description: Updated collection snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionsSnapshot"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "413":
          $ref: "#/components/responses/UserDataLimit"
    delete:
      operationId: removePlaylistTrack
      summary: Remove a track from a playlist
      responses:
        "200":
          description: Updated collection snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionsSnapshot"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/me/collections/import:
    post:
      operationId: importCollections
      summary: Merge a legacy local collection into server-authoritative data
      description: Favorites and matching playlist IDs are unioned. Missing playlist IDs receive a deterministic import ID, making retries safe for unchanged input.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CollectionsImportRequest"
      responses:
        "200":
          description: Updated collection snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionsSnapshot"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "413":
          $ref: "#/components/responses/UserDataLimit"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  parameters:
    TrackID:
      name: trackId
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 256
    PlaylistID:
      name: playlistId
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
  responses:
    UserDataLimit:
      description: The per-user persistence or entry limit was reached
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing, invalid, expired, or revoked bearer token
      headers:
        WWW-Authenticate:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadRequest:
      description: Invalid request body, candidate signature, provider, or field selection
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Track, audio file, artwork, or lyrics not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Request conflicts with current server state
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    Artwork:
      description: Track artwork image
      headers:
        Cache-Control:
          schema:
            type: string
          description: Public artwork cache policy.
        ETag:
          schema:
            type: string
          description: Content hash used for conditional requests.
        Content-Length:
          schema:
            type: integer
            format: int64
            minimum: 0
        Last-Modified:
          schema:
            type: string
          description: Artwork or audio file modification time in HTTP-date format.
        X-Content-Type-Options:
          schema:
            type: string
            const: nosniff
      content:
        image/jpeg:
          schema:
            $ref: "#/components/schemas/ArtworkBinary"
        image/png:
          schema:
            $ref: "#/components/schemas/ArtworkBinary"
        image/webp:
          schema:
            $ref: "#/components/schemas/ArtworkBinary"
        application/octet-stream:
          schema:
            $ref: "#/components/schemas/ArtworkBinary"
    ArtworkHeaders:
      description: Track artwork headers with no response body
      headers:
        Cache-Control:
          schema:
            type: string
        Content-Type:
          schema:
            type: string
        ETag:
          schema:
            type: string
        Content-Length:
          schema:
            type: integer
            format: int64
            minimum: 0
        Last-Modified:
          schema:
            type: string
          description: Artwork or audio file modification time in HTTP-date format.
        X-Content-Type-Options:
          schema:
            type: string
            const: nosniff
    AudioStream:
      description: Complete original audio file
      headers:
        Accept-Ranges:
          schema:
            type: string
          const: bytes
        ETag:
          schema:
            type: string
          description: Strong SHA-256 content validator in quoted hexadecimal form.
        Digest:
          schema:
            type: string
            example: sha-256=47DEQpj8HBSa+/TImW+5JCeuQeRkm5NMpJWZG3hSuFU=
          description: SHA-256 representation digest using the standard base64 form.
        X-Media-Version:
          schema:
            type: string
          description: Stable lowercase hexadecimal SHA-256 version of the complete audio representation.
        Content-Length:
          schema:
            type: integer
            format: int64
            minimum: 0
        Last-Modified:
          schema:
            type: string
          description: File modification time in HTTP-date format.
        Content-Disposition:
          schema:
            type: string
          description: Inline disposition with the source file name.
      content:
        audio/mpeg:
          schema:
            $ref: "#/components/schemas/AudioBinary"
        audio/flac:
          schema:
            $ref: "#/components/schemas/AudioBinary"
        audio/mp4:
          schema:
            $ref: "#/components/schemas/AudioBinary"
        audio/ogg:
          schema:
            $ref: "#/components/schemas/AudioBinary"
        audio/wav:
          schema:
            $ref: "#/components/schemas/AudioBinary"
        application/octet-stream:
          schema:
            $ref: "#/components/schemas/AudioBinary"
    PartialAudioStream:
      description: Requested byte range from the original audio file
      headers:
        Accept-Ranges:
          schema:
            type: string
          const: bytes
        ETag:
          schema:
            type: string
          description: Strong SHA-256 content validator in quoted hexadecimal form.
        Digest:
          schema:
            type: string
            example: sha-256=47DEQpj8HBSa+/TImW+5JCeuQeRkm5NMpJWZG3hSuFU=
          description: SHA-256 representation digest using the standard base64 form.
        X-Media-Version:
          schema:
            type: string
          description: Stable lowercase hexadecimal SHA-256 version of the complete audio representation.
        Content-Length:
          schema:
            type: integer
            format: int64
            minimum: 0
        Content-Range:
          schema:
            type: string
          description: Byte range included in this response.
        Last-Modified:
          schema:
            type: string
          description: File modification time in HTTP-date format.
        Content-Disposition:
          schema:
            type: string
          description: Inline disposition with the source file name.
      content:
        audio/mpeg:
          schema:
            $ref: "#/components/schemas/AudioBinary"
        audio/flac:
          schema:
            $ref: "#/components/schemas/AudioBinary"
        audio/mp4:
          schema:
            $ref: "#/components/schemas/AudioBinary"
        audio/ogg:
          schema:
            $ref: "#/components/schemas/AudioBinary"
        audio/wav:
          schema:
            $ref: "#/components/schemas/AudioBinary"
        application/octet-stream:
          schema:
            $ref: "#/components/schemas/AudioBinary"
    AudioStreamHeaders:
      description: Original audio file headers with no response body
      headers:
        Accept-Ranges:
          schema:
            type: string
          const: bytes
        ETag:
          schema:
            type: string
          description: Strong SHA-256 content validator in quoted hexadecimal form.
        Digest:
          schema:
            type: string
            example: sha-256=47DEQpj8HBSa+/TImW+5JCeuQeRkm5NMpJWZG3hSuFU=
          description: SHA-256 representation digest using the standard base64 form.
        X-Media-Version:
          schema:
            type: string
          description: Stable lowercase hexadecimal SHA-256 version of the complete audio representation.
        Content-Type:
          schema:
            type: string
        Content-Length:
          schema:
            type: integer
            format: int64
            minimum: 0
        Last-Modified:
          schema:
            type: string
          description: File modification time in HTTP-date format.
        Content-Disposition:
          schema:
            type: string
          description: Inline disposition with the source file name.
  schemas:
    CollectionsSnapshot:
      type: object
      required: [revision, favorite_track_ids, playlists]
      properties:
        revision:
          type: integer
          format: int64
          minimum: 0
        favorite_track_ids:
          type: array
          maxItems: 5000
          uniqueItems: true
          items:
            type: string
        playlists:
          type: array
          maxItems: 250
          items:
            $ref: "#/components/schemas/Playlist"
    Playlist:
      type: object
      required: [id, name, track_ids, created_at, updated_at]
      properties:
        id:
          type: string
          maxLength: 128
        name:
          type: string
          minLength: 1
          maxLength: 120
        track_ids:
          type: array
          maxItems: 5000
          uniqueItems: true
          items:
            type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    PlaylistRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
    PlaylistCreateRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        id:
          type: string
          minLength: 1
          maxLength: 128
          description: Optional client-generated ID used to make creation retries idempotent for the authenticated user.
        name:
          type: string
          minLength: 1
          maxLength: 120
    CollectionsImportRequest:
      type: object
      additionalProperties: false
      required: [favorite_track_ids, playlists]
      properties:
        revision:
          type: integer
          format: int64
          minimum: 0
          description: Optional legacy local revision. It is accepted for compatibility and not used for conflict resolution.
        favorite_track_ids:
          type: array
          maxItems: 5000
          uniqueItems: true
          items:
            type: string
        playlists:
          type: array
          maxItems: 250
          items:
            $ref: "#/components/schemas/ImportedPlaylist"
    ImportedPlaylist:
      type: object
      additionalProperties: false
      required: [name, track_ids]
      properties:
        id:
          type: string
          maxLength: 128
          description: Existing IDs merge by union; omit for a deterministic import ID.
        name:
          type: string
          minLength: 1
          maxLength: 120
        track_ids:
          type: array
          maxItems: 5000
          uniqueItems: true
          items:
            type: string
...
    CredentialsRequest:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 64
        password:
          type: string
          minLength: 12
          maxLength: 1024
          writeOnly: true
    User:
      type: object
      required: [id, username, role, created_at]
      properties:
        id:
          type: string
        username:
          type: string
        role:
          type: string
          const: admin
        created_at:
          type: string
          format: date-time
    SessionResponse:
      type: object
      required: [token, token_type, expires_at, user]
      properties:
        token:
          type: string
          writeOnly: true
        token_type:
          type: string
          const: Bearer
        expires_at:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"
    ArtworkBinary:
      type: string
      format: binary
    AudioBinary:
      type: string
      format: binary
...
    TrackListResponse:
      type: object
      required: [tracks, total]
      properties:
        tracks:
          type: array
          items:
            $ref: "#/components/schemas/Track"
        total:
          type: integer
          minimum: 0
    LibraryScanResponse:
      type: object
      required: [tracks, total, scanned_at]
      properties:
        tracks:
          type: array
          items:
            $ref: "#/components/schemas/Track"
        total:
          type: integer
          minimum: 0
        scanned_at:
          type: string
          format: date-time
    LibraryStatusResponse:
      type: object
      required: [directory, track_count, scanning, last_scanned_at, last_error]
      properties:
        directory:
          type: string
        track_count:
          type: integer
          minimum: 0
        scanning:
          type: boolean
        last_scanned_at:
          type:
            - string
            - "null"
          format: date-time
        last_error:
          type: string
    Track:
      type: object
      required: [id, title, file_name, extension, size_bytes, modified]
      properties:
        id:
          type: string
        title:
          type: string
        artist:
          type: string
        album:
          type: string
        limit:
          type: integer
          minimum: 1
          maximum: 25
          default: 15
          description: Maximum number of provider candidates to search for.
        file_name:
          type: string
        extension:
          type: string
        size_bytes:
          type: integer
          format: int64
          minimum: 0
        modified:
          type: string
          format: date-time
        artwork_url:
          type: string
          format: uri
          description: Optional explicit artwork URL. Clients may use the track artwork endpoint when omitted.
    LyricsResponse:
      type: object
      required: [track_id, encoding, content]
      properties:
        track_id:
          type: string
        encoding:
          type: string
          enum: [UTF-8, GB18030, GBK]
        content:
          type: string
          description: Lyrics normalized to UTF-8.
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string